requirements.txt
cycler==0.11.0
matplotlib==3.5.1
numpy==1.21.5
pandas==1.4.2
psutil==5.8.0
scikit_learn==1.1.1
scipy==1.7.3
sklearn_genetic==0.5.1
torch==1.11.0
